+++ Libplanet.Tools/IConfigurationService.cs
@@ -1,0 +1,9 @@
+namespace Libplanet.Tools
+{
+    public interface IConfigurationService<TConfiguration>
+    {
+        TConfiguration Load();
+
+        void Store(TConfiguration configuration);
+    }
+}